DIMPT-1: move getValue() methods to Processor and make private
